Task: refactor socket + add guards for socket

+++ backend/src/modules/room-user/roomUser.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { RoomUser } from "./roomUser.entity";
 import { Repository } from "typeorm";
@@ -11,6 +11,17 @@ export class RoomUserService {
   ) { }
 
   async createNewRoomUser(id_room: number, id_user: number): Promise<RoomUser> {
+    const roomUser: RoomUser = await this.roomUserRepository.findOne({
+      where: {
+        id_room: id_room,
+      id_user: id_user,
+      }
+    });
+
+    if(roomUser) {
+      throw new HttpException('Duplicated!', HttpStatus.CONFLICT);
+    }
+
     return await this.roomUserRepository.save({
       id_room: id_room,
       id_user: id_user,
@@ -18,6 +29,12 @@ export class RoomUserService {
   }
 
   async deleteRoomUser(id_room: number, id_user: number) {
+    const participant: RoomUser = await this.roomUserRepository.getParticipant(id_room, id_user);
+
+    if (!participant) {
+      throw new HttpException('Not found user', HttpStatus.NOT_FOUND);
+    }
+
     return await this.roomUserRepository.delete({
       id_room: id_room,
       id_user: id_user,
@@ -25,11 +42,6 @@ export class RoomUserService {
   }
 
   async getListUserOfRoom(id_room: number) {
-    return await this.roomUserRepository.createQueryBuilder('roomuser')
-      .where('id_room = :id_room', { id_room })
-      .innerJoin('roomuser.id_user', 'user')
-      .addSelect(['user.id', 'user.nickname', 'user.avatar'])
-      .orderBy('roomuser.score', 'ASC')
-      .getMany();
+    return await this.roomUserRepository.getParticipantsOfRoom(id_room);
   }
 }
